+++ src/rendering/lightingRenderPass.h
@@ -32,8 +32,8 @@ namespace phi
         vertexBuffer* _quadVbo;
         buffer* _quadEbo;
         buffer* _rtsBuffer;
-        size_t _w;
-        size_t _h;
+        float _w;
+        float _h;
         uint _quadVao;
 
     private:
@@ -44,7 +44,7 @@ namespace phi
         shader* shader;
 
     public:
-        RENDERING_API lightingRenderPass(gBufferRenderPass* gBufferPass, gl* gl, size_t w, size_t h);
+        RENDERING_API lightingRenderPass(gBufferRenderPass* gBufferPass, gl* gl, float w, float h);
         RENDERING_API ~lightingRenderPass();
 
         RENDERING_API void update();
